updata after athountication

+++ products/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect
+#from django.contrib.auth.decorators import login_required
 from .models import product
 from .forms import AddProductForm
 # Create your views here.
@@ -16,33 +17,44 @@ def product_detalis(request, pk):
                   {'Product': Product})
 
 
+# @login_required
 def product_add(request):
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES)
+    if request.user.is_authenticated and request.user.is_superuser:
+        if request.method == 'POST':
+            form = AddProductForm(request.POST, request.FILES)
 
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/products-add-successful.html')
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/products-add-successful.html')
+        else:
+            form = AddProductForm()
+        return render(request, 'products/products-add.html', {'form': form})
     else:
-        form = AddProductForm()
-    return render(request, 'products/products-add.html', {'form': form})
+        return redirect('product_list')
 
 
 def product_edit(request, pk):
-    Product = get_object_or_404(product, pk=pk)
+    if request.user.is_authenticated and request.user.is_superuser:
+        Product = get_object_or_404(product, pk=pk)
 
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES, instance=Product)
+        if request.method == 'POST':
+            form = AddProductForm(
+                request.POST, request.FILES, instance=Product)
 
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/products-add-successful.html')
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/products-add-successful.html')
+        else:
+            form = AddProductForm(instance=Product)
+        return render(request, 'products/products-add.html', {'form': form})
     else:
-        form = AddProductForm(instance=Product)
-    return render(request, 'products/products-add.html', {'form': form})
+        return redirect('product_list')
 
 
 def product_delete(request, pk):
-    Product = get_object_or_404(product, pk=pk)
-    Product.delete()
-    return redirect('product_list')
+    if request.user.is_authenticated and request.user.is_superuser:
+        Product = get_object_or_404(product, pk=pk)
+        Product.delete()
+        return redirect('product_list')
+    else:
+        return redirect('product_list')
